+++ ext_tables.sql
@@ -6,5 +6,6 @@ CREATE TABLE tx_easyshop_domain_model_product (
 	currency char(3) DEFAULT 'EUR' NOT NULL,
 	description TEXT,
 	details TEXT,
-	categories int(11) DEFAULT '0' NOT NULL
+	categories int(11) DEFAULT '0' NOT NULL,
+	images int(11) unsigned DEFAULT '0',
 );